Se colocaron verificaciones en service.py : copiar , para archivos existentes

+++ shell/commandsImpl/service.py
@@ -1,6 +1,14 @@
+import os
 def copiar(a,b):
-    a=open(a,"r")
-    b=open(b,"w")
-    b.write(a.read())
-    a.close()
-    b.close()
+    if os.path.isfile(a) :
+        a=open(a,"r")
+    else :
+        print("El archivo "+a+" no existe o no es un archivo")
+
+    if os.path.isfile(b) :
+        print("El archivo "+b+" ya existe o no es un archivo")
+    else :
+        b=open(b,"w")
+        b.write(a.read())
+        a.close()
+        b.close()
